show courseList

+++ webapp/src/main/java/com/entity/Course.java
@@ -12,7 +12,7 @@ public class Course {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "course_id")
-    private int courseId;
+    private Long courseId;
 
     @Column(name = "course_name", unique = true, nullable = false)
     private String courseName;
